requirements.txt
Flask
flask-pymongo
python-dotenv
werkzeug
deep-translator
